fixing Backend

+++ server/dto/product/product_request.go
@@ -2,7 +2,7 @@ package productdto
 
 type CreateProductRequest struct {
 	Title  string `json:"title" form:"title" gorm:"type: varchar(255)"`
-	Price  int    `json:"price" form:"price" gorm:"type: int"`
+	Price  int64  `json:"price" form:"price" gorm:"type: int"`
 	Image  string `json:"image" form:"image" gorm:"type: varchar(255)"`
 	Qty    int    `json:"qty" form:"qty" gorm:"type: int"`
 	UserId int    `json:"user_id" form:"user_id" gorm:"type: int"`
@@ -10,7 +10,7 @@ type CreateProductRequest struct {
 
 type UpdateProductRequest struct {
 	Title string `json:"title" form:"title" gorm:"type: varchar(255)"`
-	Price int    `json:"price" form:"price" gorm:"type: int"`
+	Price int64  `json:"price" form:"price" gorm:"type: int"`
 	Image string `json:"image" form:"image" gorm:"type: varchar(255)"`
 	Qty   int    `json:"qty" form:"qty" gorm:"type: int"`
 }
