+++ server/models/product.go
@@ -3,7 +3,7 @@ package models
 type Product struct {
 	ID     int                  `json:"id" gorm:"primary_key:auto_increment"`
 	Title  string               `json:"title" form:"title" gorm:"type: varchar(255)"`
-	Price  int                  `json:"price" form:"price" gorm:"type: int"`
+	Price  int64                `json:"price" form:"price" gorm:"type: int"`
 	Image  string               `json:"image" form:"image" gorm:"type: varchar(255)"`
 	Qty    int                  `json:"qty" form:"qty"`
 	UserID int                  `json:"user_id" form:"user_id"`
@@ -13,7 +13,7 @@ type Product struct {
 type ProductResponse struct {
 	ID     int                  `json:"id"`
 	Title  string               `json:"name"`
-	Price  int                  `json:"price"`
+	Price  int64                `json:"price"`
 	Image  string               `json:"image"`
 	Qty    int                  `json:"qty"`
 	UserID int                  `json:"-"`
@@ -23,7 +23,7 @@ type ProductResponse struct {
 type ProductUserResponse struct {
 	ID     int    `json:"id"`
 	Title  string `json:"name"`
-	Price  int    `json:"price"`
+	Price  int64  `json:"price"`
 	Image  string `json:"image"`
 	Qty    int    `json:"qty"`
 	UserID int    `json:"-"`
